requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
loguru==0.7.2
apscheduler==3.10.4
openpyxl==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1

